Use main.go and collapse into one file
